Move hooks to separate subdirectory and refactor hooks registration
